--- src/main/java/pe/edu/unfv/security/LoginPersonalizado.java
@@ -1,27 +1,0 @@
-package pe.edu.unfv.security;
-
-import java.io.IOException;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@Component
-public class LoginPersonalizado extends SimpleUrlAuthenticationSuccessHandler{
-
-	@Override
-	public void onAuthenticationSuccess(
-			HttpServletRequest request,
-			HttpServletResponse response,
-			FilterChain chain,
-			Authentication authentication) throws IOException, ServletException {
-		
-		response.sendRedirect("/");
-		super.onAuthenticationSuccess(request, response, chain, authentication);
-	}
-}
