--- src/main/resources/templates/acceso/login2.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-
-<head th:fragment="head(title, links)">
-	<meta charset="utf-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-	<meta name="description" content="Demo Web - 2024">
-	<meta name="author" content="Eleazar Alfredo, and Bootstrap contributors">
-	<title>Blog Template · Bootstrap v5.1</title>
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
-		integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-</head>
-
-<body>
-	<br/>
-	<div class="container">
-		<div class="card border-secondary mb-3">
-			<div class="card-header bg-secondary text-white">
-				<h1>Plantilla basica....</h1>
-			</div>
-			<div class="card-body text-secondary">
-				<div th:replace="~{layout/menu :: menu}"></div>
-				<img th:src="@{/images/pacha_mama.jpg}" width="10%" height="10%">
-			</div>
-		</div>
-	</div>
-
-</body>
-
-</html>
