--- src/main/resources/templates/acceso/login_ok.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-	<head>
-		<meta charset="utf-8" />   
-		<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" />
- 
-	</head>
-<body>
-	
-	<div class="container">
-		<div class="card border-primary mb-3">
-			<div class="card-header bg-primary text-white">
-				<h1>Login</h1>
-			</div>
-			<div class="card-body text-primary">
-				<div th:Replace="~{layout/menu :: menu}"></div>
-				
-				 <div th:if="${mensaje!=null}" th:class="'alert alert-'+${clase}+'  alert-dismissible fade show'">
-		 <span th:text="${mensaje}"> </span>
-		  <button th:if="${mensaje!=null}" type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-		 </div> 
-		 <form name="form" method="POST" th:action="@{/acceso/login}" >
-		 	<h3>Ingrese sus datos</h3>
-		 	<div class="mb-3">
-				<label for="username" class="form-label">E-Mail</label>
-				<input type="text" class="form-control" name="username" id="username" placeholder="E-Mail" required="true" autofocus="true" />
-					  
-			</div>
-		 	 <div class="mb-3">
-				<label for="password" class="form-label">Contraseña</label>
-				<input type="password" class="form-control" name="password" id="password" placeholder="Contraseña" required="true" />
-					  
-			</div>
-		 	<hr/>  
-			<!--<a href="javascript:void(0);" onclick="document.form.submit();" class="btn btn-primary">Enviar</a>-->
-			<a href="javascript:void(0);" class="btn" onclick="document.form.submit();">Enviar</a>
-		 </form>
-		 <hr/>
-		 <p>
-		 	<a th:href="@{/acceso/registro}">No tienes cuenta aún?</a>
-		 </p>
-				
-			</div>
-		</div>
-	</div>
-	
-</body>
-</html>
