--- src/main/resources/templates/home/home_.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-
-<head th:fragment="head(title, links)">
-	<meta charset="utf-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-	<meta name="description" content="Demo Web - 2024">
-	<meta name="author" content="Eleazar Alfredo, and Bootstrap contributors">
-	<title>Blog Template · Bootstrap v5.1</title>
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
-		integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-	<link th:href="@{/css/style2.css}" rel="stylesheet">
-</head>
-
-<body>
-
-	<div class="vh-100 d-flex align-items-center">
-		<div class="container">
-			<div class="row">
-				<div class="col-lg-5 col-md-6 col-sm-8 col-xs-10 mx-auto">
-					<div id="loginFrame" class="position-relative">
-						<div id="signup" class="levelDown shadow bg-white card p-3">
-							<h3 class="my-2">Signup</h3>
-							<form action="" method="" onsubmit="" class='mx-3 my-3'>
-								<div class="mb-3">
-									<label for="fullname">Full Name</label>
-									<input type="text" class="mt-2 form-control" id="fullname" name="fullname">
-								</div>
-								<div class="mb-3">
-									<label for="email">Email</label>
-									<input type="email" class="mt-2 form-control" id="email" name="email">
-								</div>
-								<div class="mb-4">
-									<label for="password">Password</label>
-									<input type="password" class="mt-2 form-control" id="password" name="password">
-								</div>
-								<div class="mb-2">
-									<button class="btn btn-lg btn-primary w-100">Login</button>
-								</div>
-							</form>
-						</div>
-						<div id="login" class="levelUp shadow bg-white card p-3">
-							<h3 class="my-2">Login</h3>
-							<form action="" method="" onsubmit="" class='mx-3 my-3'>
-								<div class="mb-3">
-									<label for="email">Email</label>
-									<input type="email" class="mt-2 form-control" id="email" name="email">
-								</div>
-								<div class="mb-2">
-									<label for="password">Password</label>
-									<input type="password" class="mt-2 form-control" id="password" name="password">
-								</div>
-								<div class="mb-2">
-									<input type="checkbox" class="me-2" for="rememeberMe"> <label
-										for="rememberMe">Remember Me</label>
-								</div>
-								<div class="mb-2">
-									<button class="btn btn-lg btn-success w-100">Login</button>
-								</div>
-								<div class="mt-4 text-center">
-									<a href="">Forgot Password?</a>
-								</div>
-							</form>
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-	</div>
-
-</body>
-
-</html>
