--- src/main/resources/templates/home/index.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-	<meta charset="utf-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-	<meta name="description" content="Demo Web - 2024">
-	<meta name="author" content="Eleazar Alfredo, and Bootstrap contributors">
-	<title>Blog Template · Bootstrap v5.1</title>
-	<link rel="stylesheet" type="text/css" th:href="@{/css/style3.css}">
-	<link href="https://fonts.googleapis.com/css?family=Poppins:600&display=swap" rel="stylesheet">
-	<link th:href="@{/css/bootstrap.min.css}" rel="stylesheet">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-</head>
-<body>
-	<img class="wave" th:src="@{/images/wave.png}">
-	<div class="container">
-		<div class="img">
-			<img th:src="@{/images/bg.svg}">
-		</div>		
-		<div class="login-content">
-			
-			<form name="form" method="post" th:action="@{/acceso/login}">
-				<img th:src="@{/images/avatar.svg}">
-				<h2 class="title">Welcome</h2>
-           		<div class="input-div one">
-           		   <div class="i">
-           		   		<i class="fas fa-user"></i>
-           		   </div>
-           		   <div class="div">           		   		
-           		   		<input type="text" class="input" name="username" id="username" required="true" autofocus="true" placeholder="E-Mail">
-           		   </div>
-           		</div>
-           		<div class="input-div pass">
-           		   <div class="i"> 
-           		    	<i class="fas fa-lock"></i>
-           		   </div>
-           		   <div class="div">           		    	
-           		    	<input type="password" class="input" name="password" id="password" required="true" autofocus="true" placeholder="Contraseña">
-            	   </div>
-            	</div>
-				<div th:if="${mensaje!=null}" th:class="'alert alert-'+${clase}+' alert-dismissible fade show'">			
-								<span th:text="${mensaje}"></span>
-								<button th:if="${mensaje!=null}" type="button" class="btn-close" data-bs-dismiss="alert" aria-label="close"></button>
-							</div>
-            	<a href="#">Forgot Password?</a>
-            	<!--<input type="submit" class="btn" value="Login">-->
-				<a href="javascript:void(0);" class="btn" onclick="document.form.submit();">Enviar</a>
-            </form>
-        </div>
-    </div>
-</body>
-</html>
