+++ src/main/resources/templates/protegido/home.html
@@ -16,7 +16,7 @@
 	<br/>
 	<div class="container">
 		<div class="card border-danger mb-3">
-			<div class="card-header bg-danger text-white">
+			<div class="card-header bg-warning text-white">
 				<h1>Home Protegido....</h1>
 			</div>
 			<div class="card-body text-danger">
