+++ src/main/resources/templates/restringido/home.html
@@ -1,5 +1,7 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org"
+	  xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+</html>
 
 <head th:fragment="head(title, links)">
 	<meta charset="utf-8" />
@@ -15,7 +17,7 @@
 	<br/>
 	<div class="container">
 		<div class="card border-danger mb-3">
-			<div class="card-header bg-danger text-white">
+			<div class="card-header bg-dark text-white">
 				<h1>Home Restringido....</h1>
 			</div>
 			<div class="card-body text-danger">
